requirements.txt
# GringoOps project requirements (pinned)
setuptools
wheel
cython
google-generativeai==0.8.5
firebase-admin==6.4.0
google-cloud-secret-manager==2.16.0
huggingface_hub==0.20.3
tiktoken==0.9.0
requests==2.32.0
openai==1.30.1
# For dashboard process management
psutil
# For testing
pytest
# Backend / Auth
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
uvicorn==0.29.0
fastapi==0.111.0
python-dotenv==1.0.1
google-cloud-firestore==2.16.0